Добавляет разделение классов и разметки
