+++ src/view/no-point-template.js
@@ -1,0 +1,3 @@
+export const createNoPointTemplate = () => {
+  return `<p class="trip-events__msg">Click New Event to create your first point</p>`;
+};
