+++ src/view/trip-cost.js
@@ -1,13 +1,5 @@
 import {createElement} from "../utils";
-
-const createTripCostTemplate = (events) => {
-  const totalPrice = events.reduce((price, event) => {
-    return price + event.price;
-  }, 0);
-  return `<p class="trip-info__cost">
-              Total: &euro;&nbsp;<span class="trip-info__cost-value">${totalPrice}</span>
-            </p>`;
-};
+import {createTripCostTemplate} from "./trip-cost-template";
 
 export default class TripCost {
   constructor(events) {
