+++ src/view/trip-events-list-template.js
@@ -1,0 +1,4 @@
+export const createEventsListTemplate = () => {
+  return `<ul class="trip-events__list">
+</ul>`;
+};
