+++ src/view/trip-menu-template.js
@@ -1,0 +1,6 @@
+export const createMenuTemplate = () => {
+  return `<nav class="trip-controls__trip-tabs  trip-tabs">
+                <a class="trip-tabs__btn  trip-tabs__btn--active" href="#">Table</a>
+               <a class="trip-tabs__btn" href="#">Stats</a>
+              </nav>`;
+};
