+++ src/view/trip-menu.js
@@ -1,11 +1,5 @@
 import {createElement} from "../utils";
-
-const createMenuTemplate = () => {
-  return `<nav class="trip-controls__trip-tabs  trip-tabs">
-                <a class="trip-tabs__btn  trip-tabs__btn--active" href="#">Table</a>
-               <a class="trip-tabs__btn" href="#">Stats</a>
-              </nav>`;
-};
+import {createMenuTemplate} from "./trip-menu-template";
 
 export default class TripMenu {
   constructor() {
